cut login fix

+++ save/entries.py
@@ -9,6 +9,7 @@ from flask_blog.models.entries import Entry
 @app.route('/')
 # ログインしていないときはログインフォームへリダイレクトし、ログインしている時はホームへ移動する
 def show_entries():
+        return redirect(url_for('/'))
 
     # データベースから全ての記事を取得して、新しい順に並べ替え
     entries = Entry.query.order_by(Entry.id.desc()).all()
@@ -19,6 +20,8 @@ def show_entries():
 # 予約投稿 データ送信はPOSTメソッド
 @app.route('/entries', methods=['POST'])
 def add_entry():
+    if not session.get('logged_in'):
+        return redirect(url_for('login'))
     
     # ログインしていればモデルインスタンスを作成
     entry = Entry(
@@ -36,13 +39,16 @@ def add_entry():
 
 @app.route('/entries/new', methods=['GET'])
 def new_entry():
+    if not session.get('logged_in'):
+        return redirect(url_for('login'))
     return render_template('entries/new.html')
 
 # url_for で渡された変数idを指定する
 # int:idとすることで、idに整数以外の値が入るとエラーになる
 @app.route('/entries/<int:id>', methods=['GET'])
 def show_entry(id):
-
+    if not session.get('logged_in'):
+        return redirect(url_for('login'))
     # 渡されたidの記事をデータベースから取得することができる
     entry = Entry.query.get(id)
     return render_template('entries/show.html', entry=entry)
@@ -50,14 +56,16 @@ def show_entry(id):
 # 編集
 @app.route('/entries/<int:id>/edit', methods=['GET'])
 def edit_entry(id):
-
+    if not session.get('logged_in'):
+        return redirect(url_for('login'))
     entry = Entry.query.get(id)
     return render_template('entries/edit.html', entry=entry)
 
 # 上書
 @app.route('/entries/<int:id>/update', methods=['POST'])
 def update_entry(id):
-
+    if not session.get('logged_in'):
+        return redirect(url_for('login'))
     entry = Entry.query.get(id)
     entry.title = request.form['title']
     entry.text = request.form['text']
@@ -68,7 +76,8 @@ def update_entry(id):
 
 @app.route('/entries/<int:id>/delete', methods=['POST'])
 def delete_entry(id):
-
+    if not session.get('logged_in'):
+        return redirect(url_for('login'))
     entry = Entry.query.get(id)
     db.session.delete(entry)
     db.session.commit()
